Issue #98: Fixing all conflicts from merging sprint 2 in to branch 98 (updating imports) and reorganizing files accordingly
