+++ frontend/src/components/Sidebar/DM-sidebar.tsx
@@ -2,7 +2,7 @@ import { useState, useEffect } from "react";
 import { Link, Navigate } from "react-router-dom";
 import { MessageCircle, Settings, User, UsersRound } from "lucide-react";
 import "./DM-sidebar.css";
-import logout from "../logout";
+import logout from "../authentication/logout";
 
 // Define a type for conversations
 interface Conversation {
