+++ frontend/src/components/Sidebar/group-sidebar.tsx
@@ -2,7 +2,7 @@ import { useState, useEffect } from "react";
 import { Link, Navigate } from "react-router-dom";
 import { MessageCircle, Settings, User, Inbox } from "lucide-react";
 import "./group-sidebar.css";
-import logout from "../logout";
+import logout from "../authentication/logout";
 
 // Define a type for channels
 interface Channel {
