+++ frontend/src/components/authentication/AuthForm.css
@@ -1,3 +1,124 @@
+.container{
+  display:flex;
+  flex-direction: column;
+  margin:auto;
+  width: 600px;
+  /* margin-top: 200px; */
+  background: #fff;
+  padding-bottom: 30px;
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+}
+
+.header{
+  display:flex;
+  flex-direction:column;
+  align-items: center;
+  gap: 9px;
+  width: 100%;
+  margin-top: 30px;
+}
+
+.text{
+  color:#142850;
+  font-size: 48px;
+  font-weight: 700;
+}
+
+.underline{
+  width:61px;
+  height:6px;
+  background: #142850;
+  border-radius: 9px;
+}
+
+.form{
+  display:flex;
+  flex-direction:column;
+}
+
+.inputs{
+  margin-top: 55px;
+  display: flex;
+  flex-direction: column;
+  gap: 25px;
+}
+
+.input{
+  display:flex;
+  align-items: center;
+  margin: auto;
+  width: 480px;
+  height: 80px;
+  background: #eaeaea;
+  border-radius: 6px;
+}
+
+.login-signup-icons{
+  margin: 0px 20px;
+  color:#142850;
+  width: 2em;
+  height: 10em;
+}
+
+.input input{
+  height: 50px;
+  width: 400px;
+  background: transparent;
+  border: none;
+  outline: none;
+  color: #797979;
+  font-size:19px;
+}
+
+.forgot-password{
+  padding-left: 62px;
+  margin-top: 27px;
+  color: #797979;
+  font-size: 18px;
+}
+
+.forgot-password span{
+  color: #00909E;
+  cursor: pointer;
+}
+
+.submit-container{
+  display: flex;
+  gap: 30px;
+  margin: 60px auto;
+}
+
+.submit{
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 220px;
+  height: 59px;
+  color: #fff;
+  background: #00909E;
+  border-radius: 50px;
+  font-size: 19px;
+  font-weight: 700;
+  cursor: pointer;
+}
+
+.submitgray{
+  background: #EAEAEA;
+  color:#676767;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 220px;
+  height: 59px;
+  border-radius: 50px;
+  font-size: 19px;
+  font-weight: 700;
+  cursor: pointer;
+}
+
 .error-message, .success-message {
   color: #21303b;
   display: flex;
